+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.25.0
+
+* QrCode and Barcode modules
+  * QrCode was requested from Codex
+  * Barcode came from an experimental script [barcode.rb](https://gist.github.com/IceDragon200/882bbed2ba576eaa96d3420069ae0680)
+
+# 0.24.0
+
+* Fixed several issues
+* Pixel order is no longer reliant on BGRA order (LE systems).
+
 # 0.23.0
 
 * Added `Minil::Image#lerp(image, delta)`